server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.27.0
